deployment fixwa
