+++ frontend/app/api/auth/update-role/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth/next";
 import { prisma } from "@/lib/prisma";
-import { authOptions } from "../[...nextauth]/route";
+import { authOptions } from "@/lib/auth-options";
 
 export async function POST(req: NextRequest) {
   try {
