+++ frontend/app/api/user/update-avatar/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/auth-options';
 import { PrismaClient } from '@prisma/client';
 
 const prisma = new PrismaClient();
